# api/openapi/spec.json
{
  "openapi": "3.0.3",
...
  "info": {
    "title": "TempConnect API",
    "version": "2.0.0",
    "description": "B2B Workforce Marketplace API — VMS, Capacity Exchange, Timesheets, RBAC, Compliance.\n\nAuthentication: Session-cookie (`tc.sid`) + `X-CSRF-Token` header for mutating requests.\n\nGet a CSRF token via `GET /api/csrf` before any POST/PATCH/DELETE.\n\n⚠️ NOTICE: This spec covers only the initial pilot routes (27 of 350+ endpoints). See docs/API_DOCUMENTATION.md for current coverage.",
    "contact": {
      "name": "TempConnect Engineering",
      "email": "tech@tempconnect.de"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    { "url": "http://localhost:3000/api", "description": "Local development" },
    { "url": "https://api.tempconnect.de/api", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tc.sid",
        "description": "HTTP-only session cookie set on login"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF token obtained from GET /api/csrf"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["company", "agency", "worker"] },
          "company_name": { "type": "string", "nullable": true },
          "phone": { "type": "string", "nullable": true },
          "postal_code": { "type": "string", "nullable": true },
          "city": { "type": "string", "nullable": true },
          "country": { "type": "string", "default": "DE" },
          "lat": { "type": "number", "nullable": true },
          "lng": { "type": "number", "nullable": true },
          "is_verified": { "type": "boolean" },
          "org_id": { "type": "string", "format": "uuid", "nullable": true },
          "plan": { "type": "string", "enum": ["FREE", "BASIS", "PLUS", "PRO", "ENTERPRISE"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [
    { "sessionCookie": [], "csrfToken": [] }
  ],
  "paths": {
...
    "/payment/checkout": {
      "post": {
        "tags": ["payment"],
        "summary": "Initiate plan upgrade checkout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["plan"],
                "properties": {
                  "plan": { "type": "string", "enum": ["BASIS", "PLUS", "PRO"] },
                  "payment_method": { "type": "string", "enum": ["demo", "stripe", "paypal"], "default": "demo" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "checkout_id": { "type": "string" },
                    "mode": { "type": "string" },
                    "redirect_url": { "type": "string", "nullable": true },
                    "stripe_session_id": { "type": "string", "nullable": true },
                    "plan": { "type": "string" },
                    "amount": { "type": "number" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid plan or payment method not available" }
        }
      }
    },
    "/payment/confirm": {
      "post": {
        "tags": ["payment"],
        "summary": "Confirm a demo checkout session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "required": ["checkout_id"], "properties": { "checkout_id": { "type": "string" } } }
            }
          }
        },
        "responses": {
          "200": { "description": "Plan activated", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "user": { "$ref": "#/components/schemas/User" } } } } } },
          "404": { "description": "Session not found" }
        }
      }
    },
...
  }
}